+++ src/index.tsx
@@ -1,40 +1,10 @@
-// ANCHOR React
-import React from 'react';
-
-// ANCHOR VideoJS
-import { VideoJsPlayerOptions, } from 'video.js';
+// ANCHOR Styles
 import './themes/video-js.css';
 import './themes/lion.css';
 
 // ANCHOR Hooks
-import { usePlayer } from './hooks/usePlayer';
+export { default as usePlayer } from './hooks/usePlayer';
 
-export interface IUncontrolledPlayerProps {
-  playerRef: React.LegacyRef<HTMLVideoElement>
-}
-
-export { usePlayer } from './hooks/usePlayer';
-
-export const UncontrolledLionPlayer = ({ playerRef }: IUncontrolledPlayerProps) => {
-  return (
-    <div data-vjs-player>
-      <video
-        ref={playerRef}
-        className="video-js vjs-theme-lion"
-      >
-        <p className="vjs-no-js">
-          To view this video please enable JavaScript, and consider upgrading to a
-          web browser that <a href="https://videojs.com/html5-video-support/" target="_blank">supports HTML5 video
-          </a>
-        </p>
-      </video>
-    </div>
-  )
-};
-
-export const LionPlayer = (props?: VideoJsPlayerOptions) => {
-  const { ref } = usePlayer({ ...props });
-
-  return <UncontrolledLionPlayer playerRef={ref} />;
-};
-
+// ANCHOR Components
+export { default as LionPlayer } from './components/LionPlayer';
+export { default as UncontrolledLionPlayer } from './components/UncontrolledLionPlayer';
